Adding builder for the StudentDTO, adding simple html page, adding Dynamic Web Module.
